Replace Broccoli build system with Parcel
I would have needed to upgrade Broccoli to keep using it anyway, so I switched to Parcel, which is easier to configure and does optimizations for me by default.

I also switched from npm to Yarn when installing Parcel and the other packages.

I changed jQuery to be imported through a listing in package.json instead of from a vendored file, which is harder to audit and update. This change has the side effect of no longer using jQuery from a CDN if possible, but I think the tiny potential slowness – 18ms on my internet – is worth the code simplicity. While I added jQuery to package.json, I also upgraded it from 2.1.4 to 3.4.1. I only use jQuery in a few places in `main.coffee` so far, in very basic ways, so there should be no incompatibilities.